requirements.txt
beautifulsoup4==4.13.2
cachetools==5.5.1
certifi==2025.1.31
charset-normalizer==3.4.1
google==3.0.0
google-api-core==2.24.1
google-api-python-client==2.160.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.66.0
httplib2==0.22.0
idna==3.10
oauthlib==3.2.2
pillow==11.1.0
proto-plus==1.26.0
protobuf==5.29.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pyparsing==3.2.1
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
soupsieve==2.6
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.3.0
